+++ src/pine/ast/hints.clj
@@ -1,4 +1,5 @@
-(ns pine.ast.hints)
+(ns pine.ast.hints
+  (:require [clojure.string :as str]))
 
 (defn- filter-relations [token candidates]
   "Given a partial token (which can be completed to a table) and all the
@@ -75,24 +76,72 @@
         ;; of duplicates
         distinct)))
 
-(defn generate [state {token :table parent :parent}]
-  (let [candidate         (-> state :tables reverse first)
-        from-alias        (state :context)
-        table-hints         (if from-alias
-
-                              ;; This is not the first table, then filter out the related tables
-                              (relation-hints state token)
-
-                              ;; This is the first table - get all the tables matching the token
-                              (table-hints state token))
-        table-hints         (if parent
-                              (filter #(= (:parent %) true) table-hints)
-                              table-hints)
+(defn generate-table-hints [state]
+  (let [{token :table parent :parent} (-> state :tables reverse first)
+        from-alias (state :context)
+        table-hints (if from-alias
+                      ;; This is not the first table, then filter out the related tables
+                      (relation-hints state token)
+                      ;; This is the first table - get all the tables matching the token
+                      (table-hints state token))
+        ;; Filter by parent if specified
+        table-hints (if parent
+                      (filter #(= (:parent %) true) table-hints)
+                      table-hints)
+        ;; Add pine expression to each hint
         add-pine-expression (fn [h] (assoc h :pine (generate-expression h)))]
     (map add-pine-expression table-hints)))
 
+(defn generate-all-column-hints
+  ([state] (generate-all-column-hints state (state :current))) ;; Overload for default `a`
+  ([state a]
+   (let [aliases (state :aliases)
+         {table :table schema :schema} (->> a (get aliases))
+         columns (if schema
+                   (get-in state [:references :schema schema :table table :columns])
+                   (get-in state [:references :table table :columns]))]
+     (for [column columns]
+       (-> column
+           (select-keys [:column])
+           (assoc :alias a))))))
+
+(defn find-relevant-columns [hints column]
+  (if column
+    (filter #(str/includes? (:column %) (:column column)) hints)
+    hints))
+
+(defn exclude-columns [hints columns]
+  (if (seq columns)
+    (filter (fn [hint]
+              (not (some #(= (:column hint) (:column %)) columns)))
+            hints)
+    hints))
+
+(defn generate-column-hints [state]
+  (let [columns (state :columns)
+        column (some-> columns reverse first)
+        a (if (seq column) (column :alias) (state :current))
+        hints (generate-all-column-hints state a)
+        type (-> state :operation :type)]
+    (cond
+      ;; If the type is :select and columns exist, filter hints using the columns
+      (and (= type :select) column)
+      (find-relevant-columns hints column)
+
+      ;; If the type is :select-partial and columns exist
+      (and (= type :select-partial) columns)
+      (exclude-columns hints columns)
+
+      ;; Otherwise, return all hints
+      :else hints)))
+
 (defn handle [state]
   (let [type (-> state :operation :type)
-        candidate (-> state :tables reverse first)
-        hints (if (and (= type :table) candidate) (generate state candidate) [])]
-    (assoc-in state [:hints] {:table hints})))
+        hints (case type
+                :table (generate-table-hints state)
+                :select (generate-column-hints state)
+                :select-partial (generate-column-hints state)
+                ;; for :where-partial, we can also use the :select hints as we need to see the columns
+                [])
+        type (case type :select-partial :select type)]
+    (assoc-in state [:hints type] (or hints []))))