+++ test/pine/hints_test.clj
@@ -13,70 +13,75 @@
       (ast/generate :test)
       :hints))
 
-(-> "company | e"
-    parser/parse
-    :result
-    (ast/generate :test)
-    ;; :hints
-    )
-
 (deftest test-hints
   (testing "Generate hints"
-    (is (= {:table [{:schema "x", :table "company"
-                     :pine "x.company"}]}
-           (gen "co")))
+    (is (= [{:schema "x", :table "company"
+             :pine "x.company"}]
+           (-> "co" gen :table)))
 
-    (is (= {:table [{:schema "y", :table "employee" :column "company_id" :parent false
-                     :pine "y.employee .company_id"}
-                    {:schema "z", :table "document", :column "company_id", :parent false,
-                     :pine "z.document .company_id"}]}
-           (gen "company | e")))
+    (is (= [{:schema "y", :table "employee" :column "company_id" :parent false
+             :pine "y.employee .company_id"}
+            {:schema "z", :table "document", :column "company_id", :parent false,
+             :pine "z.document .company_id"}]
+           (-> "company | e" gen :table)))
 
-    (is (= {:table [{:schema "x", :table "company" :column "company_id" :parent true
-                     :pine "of: x.company .company_id"}]}
-           (gen "employee | co")))
+    (is (= [{:schema "x", :table "company" :column "company_id" :parent true
+             :pine "of: x.company .company_id"}]
+           ;; (gen "employee | co")
+           (-> "employee | co" gen :table)))
 
-    (is (= {:table []}
-           (gen "company as c | s: id")))
+    (is (= []
+           (-> "company as c | s: id" gen :table)))
 
-    ;; The following shouldn't generate any hint but it does
+;; The following shouldn't generate any hint but it does
     ;;
     ;; (is (= {:table []}
     ;;        (gen "company as c")))
     )
 
   (testing "Generate hints in ambiguity"
-    (is (= {:table
-            [{:schema "z",
-              :table "document"
-              :column "employee_id"
-              :parent false
-              :pine "z.document .employee_id"}
-             {:schema "z"
-              :table "document"
-              :column "created_by"
-              :parent false
-              :pine "z.document .created_by"}]}
-           (gen "employee | doc"))))
+    (is (= [{:schema "z",
+             :table "document"
+             :column "employee_id"
+             :parent false
+             :pine "z.document .employee_id"}
+            {:schema "z"
+             :table "document"
+             :column "created_by"
+             :parent false
+             :pine "z.document .created_by"}]
+           (-> "employee | doc" gen :table))))
 
   (testing "Generate hints when direction is specified"
-    (is (= {:table
-            [{:schema "y"
-              :table "employee"
-              :column "reports_to"
-              :parent true
-              :pine "of: y.employee .reports_to"}
-             {:schema "y"
-              :table "employee"
-              :column "reports_to"
-              :parent false
-              :pine "y.employee .reports_to"}]}
-           (gen "employee | employee")))
-    (is (= {:table
-            [{:schema "y"
-              :table "employee"
-              :column "reports_to"
-              :parent true
-              :pine "of: y.employee .reports_to"}]}
-           (gen "employee | of: employee")))))
+    (is (= [{:schema "y"
+             :table "employee"
+             :column "reports_to"
+             :parent true
+             :pine "of: y.employee .reports_to"}
+            {:schema "y"
+             :table "employee"
+             :column "reports_to"
+             :parent false
+             :pine "y.employee .reports_to"}]
+           (-> "employee | employee" gen :table)))
+    (is (= [{:schema "y"
+             :table "employee"
+             :column "reports_to"
+             :parent true
+             :pine "of: y.employee .reports_to"}]
+           (-> "employee | of: employee" gen :table))))
+
+  (testing "Generate `select` hints with relevant columns"
+    (is (= []
+           (-> "x.company | s: does_not_exist" gen :select)))
+    (is (= [{:column "id" :alias "c_0"}]
+           (-> "x.company | s: i" gen :select)))
+    (is (= ["company_id" "id"] ;;  "reports_to" is not returned
+           (->> "y.employee | s: id" gen :select (map :column)))))
+
+  (testing "Generate `select-partial` hints"
+    (is (= [{:column "id" :alias "c_0"}]     (->  "company    | s:"                 gen :select)))
+    (is (= [{:column "id" :alias "c_0"}]     (->  "x.company  | s:"                 gen :select)))
+    (is (= ["reports_to"  "company_id" "id"] (->> "y.employee | s:"                 gen :select (map :column))))
+    (is (= ["reports_to"]                    (->> "y.employee | s: id, company_id," gen :select (map :column))))))