+++ CHANGELOG.md
@@ -4,6 +4,17 @@ log follows the conventions of [keepachangelog.com](http://keepachangelog.com/).
 
 ## [Unreleased]
 
+## [0.14.0] - 2025-01-07
+
+### Added
+- Support for columns e.g. hints are generated for a partial select:
+```
+company | s:
+company | s: id,
+```
+
+### Changed
+- Connection id format is `host`:`port` instead of just the `host`.
 
 ## [0.13.0] - 2024-10-25
 ### Added